Split dom-legacy into a separate markup dimensions
That way we can have separate configurations for legacy renderers
(renderToString/renderToStaticMarkup) and the modern renderToMarkup.
