+++ scripts/shared/inlinedHostConfigs.js
@@ -428,6 +428,22 @@ module.exports = [
     entryPoints: [
       'react-dom/src/server/ReactDOMLegacyServerBrowser.js', // react-dom/server.browser
       'react-dom/src/server/ReactDOMLegacyServerNode.js', // react-dom/server.node
+    ],
+    paths: [
+      'react-dom',
+      'react-dom/src/ReactDOMReactServer.js',
+      'react-dom-bindings',
+      'react-dom/src/server/ReactDOMLegacyServerImpl.js', // not an entrypoint, but only usable in *Browser and *Node files
+      'react-dom/src/server/ReactDOMLegacyServerBrowser.js', // react-dom/server.browser
+      'react-dom/src/server/ReactDOMLegacyServerNode.js', // react-dom/server.node
+      'shared/ReactDOMSharedInternals',
+    ],
+    isFlowTyped: true,
+    isServerSupported: true,
+  },
+  {
+    shortName: 'markup',
+    entryPoints: [
       'react-html/src/ReactHTMLClient.js', // react-html
       'react-html/src/ReactHTMLServer.js', // react-html/react-html.react-server
     ],
@@ -435,10 +451,6 @@ module.exports = [
       'react-dom',
       'react-dom/src/ReactDOMReactServer.js',
       'react-dom-bindings',
-      'react-server-dom-webpack',
-      'react-dom/src/server/ReactDOMLegacyServerImpl.js', // not an entrypoint, but only usable in *Browser and *Node files
-      'react-dom/src/server/ReactDOMLegacyServerBrowser.js', // react-dom/server.browser
-      'react-dom/src/server/ReactDOMLegacyServerNode.js', // react-dom/server.node
       'react-html',
       'shared/ReactDOMSharedInternals',
     ],
